Fixes filtering logic, adds fetch for markers
The fetch is just a skeleton of what needs to be done, but
as a POC it works.

Updates the filter functionality to use an object
of string|string[]|function to loop through and compare the active data
on the Marker.data object with the filter data.

E.g.
```js
const filters = {
    checkThisFunc: () => {},
    beThisValue: '',
    existInThisArray: []
}
```

This makes it easier to add new filters as the application grows

Functions should really only be used if there are side effects that
need to be added, such as on `filters.isInBounds`

Also fixes the `filters:clear` and `filters:change` event listeners

They are much more simple since they either update
an array, or assign a value to the filter property