Fixes upload issues
The issues were that the amount being "POST'ed" was more than
could be handled by either the server logic
or by sanity

The solution was to break it into chunks and create
a document for each chunk of information

This may need to be reviewed since this is a lot of documents to
create and we're limited by the sanity free tier

To get these chunked arrays into the POST
we generate a group of hidden inputs with the stringified data.

The size of the chunks may be able to be larger than 200.
500 was too much, so maybe there is a spot inbetween that works and
would lead to fewer documents